+++ spec/gemview/commands_spec.rb
@@ -157,9 +157,9 @@ RSpec.describe Gemview::Commands do
         .to receive(:choose)
         .with(
           message: "Choose a gem:",
-          choices: hash_including(birdbrain_selector => birdbrain_gem)
+          choices: a_collection_including(birdbrain_selector)
         )
-        .and_yield(birdbrain_gem)
+        .and_yield(birdbrain_selector)
 
       expect(Gemview::Terminal)
         .to receive(:choose)
@@ -263,9 +263,9 @@ RSpec.describe Gemview::Commands do
         .to receive(:choose)
         .with(
           message: "Choose a gem:",
-          choices: hash_including(rubocop_selector => rubocop_gem)
+          choices: a_collection_including(rubocop_selector)
         )
-        .and_yield(rubocop_gem)
+        .and_yield(rubocop_selector)
 
       expect(Gemview::Terminal)
         .to receive(:choose)
@@ -334,9 +334,9 @@ RSpec.describe Gemview::Commands do
         .to receive(:choose)
         .with(
           message: "Choose a gem:",
-          choices: hash_including(builder_rails_cache_selector => builder_rails_cache_gem)
+          choices: a_collection_including(builder_rails_cache_selector)
         )
-        .and_yield(builder_rails_cache_gem)
+        .and_yield(builder_rails_cache_selector)
 
       expect(Gemview::Terminal)
         .to receive(:choose)
@@ -457,9 +457,9 @@ RSpec.describe Gemview::Commands do
         .to receive(:choose)
         .with(
           message: "Choose a gem:",
-          choices: hash_including(active_model_serializers_selector => active_model_serializers_gem)
+          choices: a_collection_including(active_model_serializers_selector)
         )
-        .and_yield(active_model_serializers_gem)
+        .and_yield(active_model_serializers_selector)
 
       expect(Gemview::Terminal)
         .to receive(:choose)