+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add better support for showing the changelog
 - Add ability to sort search results by total and version downloads
 - Add ruby version to gem version output
+- Save last position in TUI selector
 
 ## [1.0.0] - 2024-12-08